Removed remark about leaving domain.
Based on review feedback. The comment

> if it causes the mesh to leave the domain being analyzed

was too vague and didn't deliver much value. One reviewer interpreted
it as having contradicted the idea of a free boundary. (The original
intent was aimed at problems on an infinite domain with a solution
that dissipates to 0 outside the domain. In these cases, boundary
elements are just there to "box in" the solution.)