Updating `main_figure30.pdf`.
This is part of the review process after journal submission. In the
script used to generate it I was just using lines for all the edges
of the refined sub-surfaces. This was an attempt to reduce the sheer
size of the SVG. In order to preserve the curvature I took a different
approach: I iterated through all sub-surfaces and only plotted
edges; tracking each edge to make sure it wasn't plotted twice (since
most edges lie on neighboring sub-surfaces).

I'd like to add that script here and may do so at a later time if
I have the bandwidth.